Move change handler to activity
