+++ ui/src/KaptureApp/components/EditorActions/connect.js
@@ -1,16 +1,16 @@
 import { connect } from 'react-redux';
 
 import { EditorActions } from './EditorActions';
-import { editor } from 'KaptureApp/store';
+import { activity, editor } from 'KaptureApp/store';
 import { actions } from 'KaptureApp/actions';
 
 const { addBarcodes } = actions.editor;
 
 const mapState = (state, props) => {
   return {
-    savePending: editor.selectSavePending(state),
-    lastSaveTime: editor.selectLastSaveTime(state),
-    saveError: editor.selectSaveError(state),
+    savePending: activity.selectSavePending(state),
+    lastSaveTime: activity.selectLastSaveTime(state),
+    saveError: activity.selectSaveError(state),
     allGridBarcodesSet: editor.selectAllGridBarcodesSet(state),
   };
 };
