+++ ui/src/KaptureApp/store/editor/slice.js
@@ -14,9 +14,6 @@ const initialState = {
       outerPadding: 2,
     },
   },
-  savePending: false,
-  saveError: null,
-  lastSaveTime: null,
   containerTypes: [],
 };
 
@@ -176,20 +173,6 @@ const editor = createSlice({
     setSettings(state, action) {
       const { settings } = action.payload;
       state.settings = Object.assign(state.settings, settings);
-    },
-    setSavePending(state, action) {
-      state.savePending = true;
-      state.saveError = null;
-      state.lastSaveTime = null;
-    },
-    setLastSaveTime(state, action) {
-      state.savePending = false;
-      state.saveError = null;
-      state.lastSaveTime = action.payload.lastSaveTime;
-    },
-    setSaveError(state, action) {
-      state.savePending = false;
-      state.saveError = action.payload.error;
     },
     setContainerTypes(state, action) {
       state.containerTypes = action.payload.containerTypes;
